Refactor constructors of importing actions
